Update nf-http-httpsendresponseentitybody.md
There seems to be a misleading text : Combining both flags in a single call to the HttpSendHttpResponse function produces undefined results.
This is documentation to HttpSendResponseEntityBody and I don't know which "both" flags are meant. The flags to HttpSendResponseEntityBody are mutually exclusive if I understand it correctly. Should it be deleted?